pyproject.toml
[tool.poetry]
name = "cron-man"
version = "0.1.0"
description = "Cron jobs for my poneys"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
click = "^8.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
